Update SignatureCreator, SignatureVerifier
